+++ example/hello-template/webapp/helloView.rsp.dart
@@ -4,8 +4,8 @@ part of hello_template;
 
 /** Template, helloView, for rendering the view. */
 void helloView(HttpConnect connect) { //4
-  final request = connect.request, response = connect.response;
-  var _v_;
+  var _cxs = new List<HttpConnect>(), request = connect.request, response = connect.response, _v_;
+
   if (!connect.isIncluded)
     response.headers.contentType = new ContentType.fromString("""text/html; charset=utf-8""");