+++ lib/src/rspc/tag.dart
@@ -6,19 +6,32 @@ part of stream_rspc;
 /** The tag execution context.
  */
 abstract class TagContext {
-  /// The parent tag, or null if not available.
-  final Tag parent;
+  /// The parent tag context, or null if this is root.
+  final TagContext parent;
   /** The output stream to generate the Dart code.
    * You can change it to have the child tag to generate the Dart code
    * to, say, a buffer.
    */
   IOSink output;
+  ///The tag
+  final Tag tag;
+  /** The map of arguments. If a tag assigns a non-null value, it means
+   * the child tags must be `var`.
+   *
+   * The key is the argument's name, while the value is the local variable's name.
+   * The local variable is used to hold the value.
+   */
+  Map<String, String> args;
+  ///Tag-specific data. A tag can store anything here.
+  var data;
   final Compiler compiler;
   ///The line number of the starting of this context
   int get line;
 
-  TagContext(this.parent, this.compiler, this.output);
+  TagContext(this.parent, this.tag, this.compiler, this.output);
 
+  ///Returns the next available name for a new local variable
+  String nextVar();
   ///The whitespace that shall be generated in front of each line
   String get pre;
   ///Indent for a new block of code. It adds two spaces to [pre].
@@ -72,7 +85,7 @@ Map<String, Tag> get tags {
   if (_tags == null) {
     _tags = new HashMap();
     for (Tag tag in [new PageTag(), new DartTag(), new HeaderTag(),
-      new IncludeTag(), new ForwardTag(),
+      new IncludeTag(), new ForwardTag(), new VarTag(),
       new ForTag(), new WhileTag(), new IfTag(), new ElseTag()])
       _tags[tag.name] = tag;
   }
@@ -82,6 +95,7 @@ Map<String, Tag> _tags;
 
 ///The page tag.
 class PageTag extends Tag {
+  @override
   void begin(TagContext tc, String data) {
     String name, desc, args, ctype;
     final attrs = ArgInfo.parse(data);
@@ -110,32 +124,40 @@ class PageTag extends Tag {
     }
     tc.compiler.setPage(name, desc, args, ctype);
   }
+  @override
   bool get hasClosing => false;
+  @override
   String get name => "page";
 }
 
 ///The dart tag.
 class DartTag extends Tag {
+  @override
   void begin(TagContext tc, String data) {
     if (!data.isEmpty)
       tc.writeln(data);
   }
+  @override
   bool get hasClosing => true;
+  @override
   String get name => "dart";
 }
 
 ///The header tag to generate HTTP response headers.
 class HeaderTag extends Tag {
+  @override
   void begin(TagContext tc, String data) {
     final attrs = ArgInfo.parse(data);
     for (final nm in attrs.keys) {
       final val = attrs[nm];
       if (val == null)
         tc.error("The $nm attribute requires a value.");
-      tc.writeln('\n${tc.pre}response.headers.add("$nm", ${toEL(val)}); //#${tc.line}');
+      tc.writeln('\n${tc.pre}response.headers.add("$nm", ${toEL(val)}); //header#${tc.line}');
     }
   }
+  @override
   bool get hasClosing => false;
+  @override
   String get name => "header";
 }
 
@@ -151,16 +173,32 @@ class HeaderTag extends Tag {
  * placed inside others.
  */
 class IncludeTag extends Tag {
+  @override
   void begin(TagContext tc, String data) {
-    final argInfo = new ArgInfo(tc, data);
+    tc.data = new ArgInfo(tc, data);
+    tc.args = new LinkedHashMap(); //order is important
+  }
+  @override
+  void end(TagContext tc) {
+    final argInfo = tc.data;
+    final args = _mergeArgs(argInfo.args, tc.args);
     if (argInfo.isID)
-      tc.compiler.include(argInfo.first, argInfo.args, tc.line);
+      tc.compiler.include(argInfo.first, args, tc.line);
     else
-      tc.compiler.includeUri(argInfo.first, argInfo.args, tc.line);
+      tc.compiler.includeUri(argInfo.first, args, tc.line);
   }
 
-  bool get hasClosing => false;
+  @override
+  bool get hasClosing => true;
+  @override
   String get name => "include";
+}
+///merge arguments
+Map<String, dynamic> _mergeArgs(Map<String, dynamic> dst, Map<String, String> src) {
+  if (src != null)
+    for (final nm in src.keys)
+      dst[nm] = "[=${src[nm]}.toString()]";
+  return dst;
 }
 
 /** The forward tag. There are two formats:
@@ -176,20 +214,57 @@ class IncludeTag extends Tag {
  * Unlike the include tag, it can be placed inside another tags.
  */
 class ForwardTag extends Tag {
+  @override
   void begin(TagContext tc, String data) {
-    final argInfo = new ArgInfo(tc, data);
-    if (argInfo.isID)
-      tc.compiler.forward(argInfo.first, argInfo.args, tc.line);
-    else
-      tc.compiler.forwardUri(argInfo.first, argInfo.args, tc.line);
+    tc.data = new ArgInfo(tc, data);
+    tc.args = new LinkedHashMap(); //order is important
   }
-
-  bool get hasClosing => false;
+  @override
+  void end(TagContext tc) {
+    final argInfo = tc.data;
+    final args = _mergeArgs(argInfo.args, tc.args);
+    if (argInfo.isID)
+      tc.compiler.forward(argInfo.first, args, tc.line);
+    else
+      tc.compiler.forwardUri(argInfo.first, args, tc.line);
+  }
+  @override
+  bool get hasClosing => true;
+  @override
   String get name => "forward";
+}
+
+/** The var tag. It defines a variable or an argument passed to [IncludeTag]
+ * or [ForwardTag]
+ */
+class VarTag extends Tag {
+  @override
+  void begin(TagContext tc, String data) {
+    final argInfo = new ArgInfo(tc, data, strFirst: false);
+    final parentArgs = tc.parent.args;
+    var varnm = tc.data
+      = parentArgs != null ? (parentArgs[argInfo.first] = tc.nextVar()): argInfo.first;
+
+    tc.writeln("\n${tc.pre}var $varnm = new StringBuffer(); _cxs.add(connect); //var#${tc.line}\n"
+      "${tc.pre}connect = new HttpConnect.buffer(connect, $varnm); response = connect.response;");
+  }
+  @override
+  void end(TagContext tc) {
+    tc.writeln("\n${tc.pre}connect = _cxs.removeLast(); response = connect.response;");
+    if (tc.parent.args == null) {
+      String varnm = tc.data;
+      tc.writeln("${tc.pre}$varnm = $varnm.toString();");
+    }
+  }
+  @override
+  bool get hasClosing => true;
+  @override
+  String get name => "var";
 }
 
 ///A skeletal class for implementing control tags, such as [IfTag] and [WhileTag].
 abstract class ControlTag extends Tag {
+  @override
   void begin(TagContext tc, String data) {
     if (data.isEmpty)
       tc.error("The $name tag requires a condition");
@@ -201,17 +276,21 @@ abstract class ControlTag extends Tag {
       beg = needsVar ? "(var ": "(";
       end = ")";
     }
-    tc.writeln("\n${tc.pre}$control $beg$data$end { //#${tc.line}");
+    tc.writeln("\n${tc.pre}$control $beg$data$end { //$name#${tc.line}");
     tc.indent();
   }
+  @override
   void end(TagContext tc) {
     tc.unindent();
     tc.writeln("${tc.pre}} //$control");
   }
+  @override
   bool get hasClosing => true;
   ///The name of the control, such as `if` and `while`. Default: [name].
+  @override
   String get control => name;
   ///Whether `var` is required in front of the condition
+  @override
   bool get needsVar => false;
 }
 
@@ -224,7 +303,9 @@ abstract class ControlTag extends Tag {
  *     [/for]
  */
 class ForTag extends ControlTag {
+  @override
   String get name => "for";
+  @override
   bool get needsVar => true;
 }
 
@@ -234,6 +315,7 @@ class ForTag extends ControlTag {
  *     [/while]
  */
 class WhileTag extends ControlTag {
+  @override
   String get name => "while";
 }
 
@@ -243,6 +325,7 @@ class WhileTag extends ControlTag {
  *      [/if]
  */
 class IfTag extends  ControlTag {
+  @override
   String get name => "if";
 }
 
@@ -255,14 +338,15 @@ class IfTag extends  ControlTag {
  */
 class ElseTag extends Tag {
   //The implementation is a bit tricky: it pretends to be an tag without the closing.
+  @override
   void begin(TagContext tc, String data) {
-    if (!(tc.parent is IfTag))
+    if (!(tc.parent.tag is IfTag))
       tc.error("Unexpected else tag");
 
     tc.unindent();
 
     if (data.isEmpty) {
-      tc.writeln("\n${tc.pre}} else { //#${tc.line}");
+      tc.writeln("\n${tc.pre}} else { //else#${tc.line}");
     } else {
       String cond;
       if (data.length < 4 || !data.startsWith("if")
@@ -277,11 +361,13 @@ class ElseTag extends Tag {
         beg = "(";
         end = ")";
       }
-      tc.writeln("\n${tc.pre}} else if $beg$cond$end { //#${tc.line}");
+      tc.writeln("\n${tc.pre}} else if $beg$cond$end { //else#${tc.line}");
     }
 
     tc.indent();
   }
+  @override
   bool get hasClosing => false;
+  @override
   String get name => "else";
 }